data jenkins

+++ Jenkinsfile
@@ -1,10 +1,10 @@
 pipeline{
-    agent{label "principal"}
+    agent any
     parameters{
         string(name: 'name_container', defaultValue: 'sitio_web', description: 'Nombre del container')
         string(name: 'name_imagen', defaultValue: 'php', description: 'Nombre de la imagen')
         string(name: 'tag_imagen', defaultValue: '7.4-apache', description: 'etiqueta y/o version de la imagen')
-        string(name: 'puerto_imagen', defaultValue: '8090', description: 'puerto de la imagen')
+        string(name: 'puerto_imagen', defaultValue: '80', description: 'puerto de la imagen')
     }
 
 
@@ -12,14 +12,24 @@ pipeline{
 
         stage('detener/limpiar'){
             steps{
-                sh '''
-                #sudo service nginx restart 
-                sudo docker compose down
-                sudo docker stop ${name_container}
-                sudo docker rm ${name_container}
-                #sudo docker system prune -f
-                #sudo docker images purge
-                '''
+                script{
+                def docker_running = sh(returnStatus: true, script: 'sudo docker ps')
+                if (docker_running) {
+            			sh '''
+                        #sudo service nginx restart 
+                        #sudo docker compose down
+                        sudo docker stop ${name_container}
+                        sudo docker rm ${name_container}
+                        sudo docker system prune -f
+                        sudo docker images purge
+                        '''
+        		    } else {
+            			sh '''
+                        echo 'no esta'
+                        '''
+        		    }
+                }
+                
             }
         }
